Moved models to shared project
