+++ Server/src/main/java/com/bbd/BeanServer/controller/PostController.java
@@ -1,9 +1,10 @@
 package com.bbd.BeanServer.controller;
 
 import com.bbd.BeanServer.assembler.ModelAssembler;
-import com.bbd.BeanServer.model.Post;
 import com.bbd.BeanServer.repository.PostRepository;
 import com.bbd.BeanServer.service.PostService;
+import com.bbd.shared.models.Post;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
