+++ Server/src/main/java/com/bbd/BeanServer/service/PostService.java
@@ -4,8 +4,8 @@ package com.bbd.BeanServer.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.bbd.BeanServer.model.Post;
 import com.bbd.BeanServer.repository.PostRepository;
+import com.bbd.shared.models.Post;
 
 @Service
 public class PostService {
