+++ shared_resources/src/main/java/com/bbd/shared/models/FavoriteBean.java
@@ -1,11 +1,10 @@
-package com.bbd.BeanServer.model;
+package com.bbd.shared.models;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
