+++ shared_resources/src/main/java/com/bbd/shared/models/Tag.java
@@ -1,4 +1,4 @@
-package com.bbd.BeanServer.model;
+package com.bbd.shared.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,18 +9,14 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
-import java.sql.Timestamp;
-
 @Entity
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Reaction {
+public class Tag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int reaction_id;
-    int user_id;
-    int reaction_type_id;
-    Timestamp date_of_reaction;
+    int tag_id;
+    String tag_name;
 }
